cluster k8s rearranging
