+++ src/pkg/k8s/namespace.go
@@ -73,31 +73,6 @@ func (k *K8s) DeleteNamespace(ctx context.Context, name string) error {
 	}
 }
 
-// NewZarfManagedNamespace returns a corev1.Namespace with Zarf-managed labels
-func NewZarfManagedNamespace(name string) *corev1.Namespace {
-	namespace := &corev1.Namespace{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: corev1.SchemeGroupVersion.String(),
-			Kind:       "Namespace",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-		},
-	}
-	namespace.Labels = AdoptZarfManagedLabels(namespace.Labels)
-	return namespace
-}
-
-// AdoptZarfManagedLabels adds & deletes the necessary labels that signal to Zarf it should manage a namespace
-func AdoptZarfManagedLabels(labels map[string]string) map[string]string {
-	if labels == nil {
-		labels = make(map[string]string)
-	}
-	labels[ZarfManagedByLabel] = "zarf"
-	delete(labels, AgentLabel)
-	return labels
-}
-
 // IsInitialNamespace returns true if the given namespace name is an initial k8s namespace: https://kubernetes.io/docs/concepts/overview/working-with-objects/namespaces/#initial-namespaces
 func (k *K8s) IsInitialNamespace(name string) bool {
 	if name == "default" {
